More tests, error for auth challenge failure
